Laser Emitter, Mirror and King exist
All pieces have collision.  Collision Layer 1 is reserved for reflective surfaces and Collision Layer 2 is reserved for opaque surfaces.  Layers 3 and 4 are currently reserved for the laser and world boundaries, though they are currently unused.  The laser emitter piece has a raycast that can detect if it collides with a surface.  Next step is to have the raycast be able to determine if the hit surface is reflective or non-reflective and either create a new raycast or destroy the piece appropriately