Reflection works
moved code for firing the laser from the emitter to the raycast itself.  Each mirror has its own raycast, and if the laser hits a mirror, it passes the appropriate parameters to the mirror's raycast, which fires as a laser.  This laser can then destroy objects like the original laser