Added WorldBoundaries to Board
Collision Layer 4 is reserved for boundaries to be placed around every board.  This allows for a beam segment that misses all of the pieces and goes off the board to be rendered.  This also helps isolate multiple boards in 2D space so we can fake 3D with them in future