Players alternate turns
The behavior of the "End Turn" button has changed.  The first time it is pressed, P1's laser is fired and shown on screen.  When it is pressed again, P1's laser turns off, starting P2's turn.  The next time it is pressed, P2's laser is fired and shown on screen.  Pressing it a fourth time turns off P2's laser.  This behavior then loops.  The main scene tracks the current player