Double Mirror Piece
Added a piece that is mirrored on both sides.  Note that with 45 degree rotations, it's possible for the laser to hit the piece parallel to the surfaces of both mirrors (i.e. go between them).  Currently the piece is set up so that if that happens, the laser will just go through unobstructed.  That behavior can be changed if needed though!